// client/stocklinx app/src/redux/company/constant.ts
export enum companyConst {
  FETCH_COMPANIES_REQUEST = "FETCH_COMPANIES_REQUEST",
  FETCH_COMPANIES_SUCCESS = "FETCH_COMPANIES_SUCCESS",
  FETCH_COMPANIES_FAILURE = "FETCH_COMPANIES_FAILURE",
  FETCH_COMPANY_REQUEST = "FETCH_COMPANY_REQUEST",
  FETCH_COMPANY_SUCCESS = "FETCH_COMPANY_SUCCESS",
  FETCH_COMPANY_FAILURE = "FETCH_COMPANY_FAILURE",
  FETCH_COMPANY_COUNTS_REQUEST = "FETCH_COMPANY_COUNTS_REQUEST",
  FETCH_COMPANY_COUNTS_SUCCESS = "FETCH_COMPANY_COUNTS_SUCCESS",
  FETCH_COMPANY_COUNTS_FAILURE = "FETCH_COMPANY_COUNTS_FAILURE",
  CREATE_COMPANY_REQUEST = "CREATE_COMPANY_REQUEST",
  CREATE_COMPANY_SUCCESS = "CREATE_COMPANY_SUCCESS",
  CREATE_COMPANY_FAILURE = "CREATE_COMPANY_FAILURE",
  CREATE_RANGE_COMPANY_REQUEST = "CREATE_RANGE_COMPANY_REQUEST",
  CREATE_RANGE_COMPANY_SUCCESS = "CREATE_RANGE_COMPANY_SUCCESS",
  CREATE_RANGE_COMPANY_FAILURE = "CREATE_RANGE_COMPANY_FAILURE",
  UPDATE_COMPANY_REQUEST = "UPDATE_COMPANY_REQUEST",
  UPDATE_COMPANY_SUCCESS = "UPDATE_COMPANY_SUCCESS",
  UPDATE_COMPANY_FAILURE = "UPDATE_COMPANY_FAILURE",
  REMOVE_COMPANY_REQUEST = "REMOVE_COMPANY_REQUEST",
  REMOVE_COMPANY_SUCCESS = "REMOVE_COMPANY_SUCCESS",
  REMOVE_COMPANY_FAILURE = "REMOVE_COMPANY_FAILURE",
  REMOVE_RANGE_COMPANY_REQUEST = "REMOVE_RANGE_COMPANY_REQUEST",
  REMOVE_RANGE_COMPANY_SUCCESS = "REMOVE_RANGE_COMPANY_SUCCESS",
  REMOVE_RANGE_COMPANY_FAILURE = "REMOVE_RANGE_COMPANY_FAILURE",
  FILTER_COMPANIES_REQUEST = "FILTER_COMPANIES_REQUEST",
  FILTER_COMPANIES_SUCCESS = "FILTER_COMPANIES_SUCCESS",
  FILTER_COMPANIES_FAILURE = "FILTER_COMPANIES_FAILURE",
  SET_COMPANY = "SET_COMPANY",
  CLEAR_COMPANY = "CLEAR_COMPANY",
  SET_COMPANIES = "SET_COMPANIES",
  CLEAR_COMPANIES = "CLEAR_COMPANIES",
}